pyproject.toml
[project]
name = "optimizing-agent-fidelity-ma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic[bedrock]>=0.75.0",
    "dotenv>=0.9.9",
    "fastapi>=0.127.0",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "openai>=2.14.0",
    "pandas>=3.0.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
    "qwen-agent>=0.0.31",
    "scikit-learn>=1.8.0",
    "tenacity>=9.0.0",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
